requirements.txt
streamlit==1.51.0
python-dotenv==1.2.1
openai==2.7.1
requests==2.32.5
Pillow==12.0.0
pygame==2.6.1
SpeechRecognition
PyAudio

# Useful extras you may want later:
# pandas==2.3.3
# pyarrow==21.0.0